some code refactoring

+++ lib/data/repositories/address/address_repository.dart
@@ -6,8 +6,10 @@ import '../authentication/authentication_repository.dart';
 
 class AddressRepository extends GetxController {
   static AddressRepository get instance => Get.find();
-  final supabase = Supabase.instance.client;
+  final _db = Supabase.instance.client;
+  final String _table = 'addresses';
 
+  /// Extraire les addresses utilisateur; Ajouter addresse; Mettre à jour adresse; Supprimer adresse
   Future<List<AddressModel>> fetchUserAddresses() async {
     try {
       final userId = AuthenticationRepository.instance.authUser!.id;
@@ -15,8 +17,7 @@ class AddressRepository extends GetxController {
         throw ('Unable to find user information. Try again in few minutes');
       }
 
-      final response =
-          await supabase.from('addresses').select().eq('user_id', userId);
+      final response = await _db.from(_table).select().eq('user_id', userId);
 
       // response is List<dynamic>
       final data = response as List<dynamic>;
@@ -28,26 +29,12 @@ class AddressRepository extends GetxController {
     }
   }
 
-  Future<void> updateSelectedField(String addressId, bool selected) async {
-    try {
-      final userId = AuthenticationRepository.instance.authUser!.id;
-
-      await supabase
-          .from('addresses')
-          .update({'selected_address': selected})
-          .eq('id', addressId)
-          .eq('user_id', userId);
-    } catch (e) {
-      throw 'Something went wrong while updating address information, try again later';
-    }
-  }
-
   Future<String> addAddress(AddressModel address) async {
     try {
       final userId = AuthenticationRepository.instance.authUser!.id;
 
-      final response = await supabase
-          .from('addresses')
+      final response = await _db
+          .from(_table)
           .insert({
             ...address.toJson(),
             'user_id': userId,
@@ -62,7 +49,20 @@ class AddressRepository extends GetxController {
     }
   }
 
-  /// Delete an address by ID
+  Future<void> selectOtherAddress(String addressId, bool selected) async {
+    try {
+      final userId = AuthenticationRepository.instance.authUser!.id;
+
+      await _db
+          .from(_table)
+          .update({'selected_address': selected})
+          .eq('id', addressId)
+          .eq('user_id', userId);
+    } catch (e) {
+      throw 'Something went wrong while updating address information, try again later';
+    }
+  }
+
   Future<void> deleteAddress(String addressId) async {
     try {
       final userId = AuthenticationRepository.instance.authUser!.id;
@@ -70,11 +70,7 @@ class AddressRepository extends GetxController {
         throw ('Unable to find user information. Try again in few minutes');
       }
 
-      await supabase
-          .from('addresses')
-          .delete()
-          .eq('id', addressId)
-          .eq('user_id', userId);
+      await _db.from(_table).delete().eq('id', addressId).eq('user_id', userId);
     } catch (e) {
       throw 'Something went wrong while deleting address information, try again later';
     }
